Create Leap Year question
Leap years. Write a function:
bool leap_year(int year)
that tests whether a year is a leap year: that is, a year with 366
days. Leap years are necessary to keep the calendar
synchronized with the sun because the earth revolves around
the sun once every 365.25 days. Actually, that figure is not
entirely precise, and for all dates after 1582 the Gregorian
correction applies. Usually, years that are divisible by 4 are leap
years, for example 1996. However, years that are divisible by
100 (for example, 1900) are not leap years, but years that are divisible by 400 are leap
years (for example, 2000).